// src/stats/types/assist-tracker.ts
export interface AssistTrackerParams {
  college?: string;
  conference?: string;
  country?: string;
  dateFrom?: string;
  dateTo?: string;
  division?: string;
  draftPick?: string;
  draftYear?: number;
  gameScope?: string;
  height?: string;
  lastNGames?: number;
  leagueID?: string;
  location?: string;
  month?: number;
  opponentTeamID?: number;
  outcome?: string;
  pORound?: number;
  perMode?: string;
  playerExperience?: string;
  playerPosition?: string;
  season?: string;
  seasonSegment?: string;
  seasonType?: string;
  starterBench?: string;
  teamID?: number;
  vsConference?: string;
  vsDivision?: string;
  weight?: string;
}

export interface AssistTrackerResponse {
}